wishlist ready

+++ src/wishlist/wishlist.service.ts
@@ -1,57 +1,68 @@
-import { Injectable } from '@nestjs/common';
-import { DeleteResult, FindManyOptions, FindOneOptions, Repository } from 'typeorm';
+import { ForbiddenException, HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { DeleteResult, FindManyOptions, FindOneOptions, In, Repository } from 'typeorm';
 import { WishList } from './entities/wishlist.entity';
 import { InjectRepository } from '@nestjs/typeorm';
-import { User } from 'src/users/entities/user.entity';
 import { CreateWishlistDto } from './dto/create-wishlist.dto';
-import { UserService } from 'src/users/user.service';
+import { UserPublicProfileResponseDto } from 'src/users/dto/public-profile-response.dto';
 import { WishService } from 'src/wishes/wishes.service';
-import { Wish } from 'src/wishes/entities/wishes.entity';
+import { UpdateWishlistDto } from './dto/update-wishlist.dto';
 
 @Injectable()
 export class WishlistService {
   constructor(
     @InjectRepository(WishList)
     private readonly wishlistsRepo: Repository<WishList>,
-    @InjectRepository(Wish)
-    private readonly wishRepo: Repository<Wish>,
-    private readonly userService: UserService
+    private readonly wishService: WishService,
   ) { }
 
   async findOne(query: FindOneOptions<WishList>): Promise<WishList> {
-    const wishlist = await this.wishlistsRepo.findOne(query);
-    console.log(wishlist)
-    return wishlist
+    return await this.wishlistsRepo.findOne(query);
   }
 
   async find(query: FindManyOptions<WishList>): Promise<WishList[]> {
-    const wishlist = await this.wishlistsRepo.find(query);
-    return wishlist
+    return await this.wishlistsRepo.find(query);
   }
 
-  async create(createWishlistDto: CreateWishlistDto, userId: number) {
+  async create(createWishlistDto: CreateWishlistDto, user: UserPublicProfileResponseDto): Promise<WishList> {
+    const { itemsId: items } = createWishlistDto;
 
-    const owner = await this.userService.findOne(
-      { where: { id: userId } }
-    );
+    const wishes = await this.wishService.find({
+      where: { id: In(items) }
+    });
+
+    if (!wishes.length) {
+      throw new Error('No wishes found for the provided IDs');
+    }
 
     const wishlist = this.wishlistsRepo.create({
-      ...createWishlistDto, owner
-    })
-    const wishes = await this.wishRepo.find({
-      where: { owner: { id: userId } }
+      ...createWishlistDto,
+      owner: user,
+      items: [...wishes]
     });
-    wishlist.items = wishes;
 
-    return this.wishlistsRepo.save(wishlist);
-
+    try {
+      return await this.wishlistsRepo.save(wishlist);
+    } catch (error) {
+      throw new Error('Error saving wishlist: ' + error.message);
+    }
   }
 
-  //   async getWishlist(id: number): Promise<WishList> {
-  //     return this.wishlistsRepo.findOne(id, { relations: ['items'] });
-  // }
+  async updateOne(query: FindOneOptions<WishList>, body: UpdateWishlistDto, userId: number): Promise<WishList> {
+    try {
+      const entity = await this.findOne(query);
+      if (entity.owner.id !== userId) throw new ForbiddenException('Редактирование чужих вишлистов запрещено');
+      return this.wishlistsRepo.save({ ...entity, ...body });
+    } catch (error) {
+      throw new HttpException("Ошибка при обновлении", HttpStatus.BAD_REQUEST);
+    }
+  }
 
-  async removeOne(query: Partial<WishList>): Promise<DeleteResult> {
-    return await this.wishlistsRepo.delete(query)
+  async removeOne(query: Partial<WishList>, userId: number): Promise<DeleteResult> {
+    const { id } = query;
+    const wish = await this.findOne({
+      where: { id }, relations: ['owner']
+    })
+    if (wish.owner.id !== userId) throw new ForbiddenException('Удаление чужих вишлистов запрещено');
+    return await this.wishlistsRepo.delete(query);
   }
 }
